Moved to 40 bits addresses.

+++ src/main/scala/db/file/writer/ByteHandler.scala
@@ -5,6 +5,6 @@ import db.util.DatabaseUtils.ByteSize
 
 class ByteHandler extends DataHandler {
   override val Size: Int = ByteSize
-  override def write(context: FileContext, offset: Int, value: Long): Unit = context.writeByte(offset, value.toInt)
-  override def read(context: FileContext, offset: Int): Long = context.readByte(offset)
+  override def write(context: FileContext, offset: Long, value: Long): Unit = context.writeByte(offset, value.toInt)
+  override def read(context: FileContext, offset: Long): Long = context.readByte(offset)
 }
