+++ src/main/scala/db/file/writer/DataHandler.scala
@@ -5,7 +5,7 @@ import db.file.FileContext
 abstract class DataHandler {
 
   val Size: Int
-  def write(context: FileContext, offset: Int, value: Long): Unit
-  def read(context: FileContext, offset: Int): Long
+  def write(context: FileContext, offset: Long, value: Long): Unit
+  def read(context: FileContext, offset: Long): Long
 
 }
