+++ src/main/scala/db/file/writer/IntHandler.scala
@@ -5,6 +5,6 @@ import db.util.DatabaseUtils.IntSize
 
 class IntHandler extends DataHandler {
   override val Size: Int = IntSize
-  override def write(context: FileContext, offset: Int, value: Long): Unit = context.writeInt(offset, value.toInt)
-  override def read(context: FileContext, offset: Int): Long = context.readInt(offset)
+  override def write(context: FileContext, offset: Long, value: Long): Unit = context.writeInt(offset, value.toInt)
+  override def read(context: FileContext, offset: Long): Long = context.readInt(offset)
 }
